refactor: time_used handling to use float and NUMERIC types across models and scripts; update backfill script for consistency with shared time-parser utility.

+++ app/schemas.py
@@ -65,7 +65,7 @@ class DrillPositionResponse(BaseModel):
     initial_eval: float
     eval_swing: float
     created_at: datetime
-    time_used: Optional[int] = None
+    time_used: Optional[float] = None
     hero_result: str
     result_reason: str
     time_control: str
